+++ ProgramFunctions/Help.fnc.php
@@ -1,0 +1,214 @@
+<?php
+/**
+ * Help functions
+ * Mainly used in Help.php, Help_en.php & Bottom.php
+ *
+ * @package RosarioSIS
+ * @subpackage ProgramFunctions
+ */
+
+/**
+ * Gettext translation function for Help texts.
+ * Registers domain on first run.
+ *
+ * @since  3.9
+ * @since  4.3 Moved from Help_en.php to ProgramFunctions/Help.fnc.php
+ *
+ * @param  string $text      Text to translate.
+ * @param  string $domain    Gettext domain, defaults to 'help'. For add-ons, use the module / plugin name / folder.
+ * @return string Translated help text.
+ */
+function _help( $text, $domain = 'help' )
+{
+	global $LocalePath;
+
+	static $domains_bound = array();
+
+	$addon = $domain;
+
+	if ( $domain !== 'help'
+		&& ! mb_strpos( $domain, 'help' ) )
+	{
+		$domain .= '_help';
+	}
+
+	if ( empty( $domains_bound[$domain] ) )
+	{
+		$locale_path = $LocalePath;
+
+		if ( $addon !== 'help' )
+		{
+			$locale_path = 'modules/' . $addon . '/locale';
+
+			if ( ! file_exists( $locale_path ) )
+			{
+				// Is plugin?
+				$locale_path = 'plugins/' . $addon . '/locale';
+
+				if ( ! file_exists( $locale_path ) )
+				{
+					return $text;
+				}
+			}
+		}
+
+		// Binds the messages domain to the locale folder.
+		bindtextdomain( $domain, $locale_path );
+
+		if ( function_exists( '_bindtextdomain' ) )
+		{
+			// Correctly bind domain when MoTranslator is in use.
+			_bindtextdomain( $domain, $locale_path );
+		}
+
+		// Ensures text returned is utf-8, quite often this is iso-8859-1 by default.
+		bind_textdomain_codeset( $domain, 'UTF-8' );
+
+		$domains_bound[$domain] = true;
+	}
+
+	return dgettext( $domain, $text );
+}
+
+
+/**
+ * Load Help
+ * English, translated (locale) & non core modules help texts
+ *
+ * @since 4.3
+ *
+ * @param boolean $force Force loading help if was already loaded.
+ *
+ * @return array $help
+ */
+function HelpLoad( $force = false )
+{
+	static $help_loaded = false;
+
+	global $RosarioModules,
+		$RosarioCoreModules,
+		$help;
+
+	if ( $help_loaded
+		&& ! $force )
+	{
+		return $help;
+	}
+
+	require_once 'Help_en.php';
+
+	// Add help for non-core modules.
+	$non_core_modules = array_diff( array_keys( $RosarioModules ), $RosarioCoreModules );
+
+	$help_english = 'Help_en.php';
+
+	// @deprecated since 3.9 use help text domain: help.po Gettext files.
+	$help_translated = 'Help_' . substr( $locale, 0, 2 ) . '.php';
+
+	foreach ( (array) $non_core_modules as $non_core_module )
+	{
+		$non_core_dir = 'modules/' . $non_core_module . '/';
+
+		if ( file_exists( $non_core_dir . $help_translated ) ) // FJ translated help.
+		{
+			require_once $non_core_dir . $help_translated;
+		}
+		elseif ( file_exists( $non_core_dir . $help_english ) )
+		{
+			require_once $non_core_dir . $help_english;
+		}
+	}
+
+	$help_loaded = true;
+
+	return $help;
+}
+
+
+/**
+ * Get Help text for program (modname)
+ * Defaults to 'default' and formats Help:
+ * - Replace 'your child' with 'you' & 'your child\'s' with 'your' for students
+ * - Replace 'RosarioSIS' with configured app name.
+ *
+ * @since 4.3
+ *
+ * @uses GetHelpTextRaw()
+ *
+ * @param string $modname Program, typically $_REQUEST['modname'].
+ *
+ * @return string Help text.
+ */
+function GetHelpText( $modname )
+{
+	$help_text = GetHelpTextRaw( $modname );
+
+	// Get default help text.
+	if ( empty( $help_text ) )
+	{
+		$help_text = $help['default'];
+	}
+
+	if ( User( 'PROFILE' ) === 'student' )
+	{
+		$help_text = str_replace(
+			array( 'your child\'s', 'your child' ),
+			array( 'your', 'yourself' ),
+			$help_text
+		);
+	}
+
+	// Replace RosarioSIS with configured app name.
+	$help_text = str_replace( 'RosarioSIS', Config( 'NAME' ), $help_text );
+
+	return $help_text;
+}
+
+
+
+/**
+ * Get raw Help text for program (modname)
+ * No default and no formatting.
+ *
+ * @since 4.3
+ *
+ * @uses HelpLoad()
+ *
+ * @example $program_has_help_text = GetHelpTextRaw( $_REQUEST['modname'] );
+ *
+ * @param string $modname Program, typically $_REQUEST['modname'].
+ *
+ * @return string Help text.
+ */
+function GetHelpTextRaw( $modname )
+{
+	$help = HelpLoad();
+
+	if ( empty( $modname ) )
+	{
+		return '';
+	}
+
+	$help_text = '';
+
+	if ( ! empty( $help[ $modname ] ) )
+	{
+		$help_text = $help[ $modname ];
+	}
+	else
+	{
+		foreach ( (array) $help as $program => $help_txt )
+		{
+			// FJ fix bug URL Modules.php?modname=Student_Billing/Statements.php&_ROSARIO_PDF.
+			if ( ( mb_strpos( $program, $modname ) === 0
+					&& mb_strpos( $_SERVER['QUERY_STRING'], $program ) === 21 ) )
+			{
+				$help_text = $help_txt;
+
+				break;
+			}
+		}
+	}
+
+	return $help_text;
+}